+++ packages/react/src/highlight-indicator/index.ts
@@ -1,3 +1,5 @@
-export { useHighlightIndicator } from './useHighlightIndicator'
+export { useHighlightIndicator, useHighlightStore } from './useHighlightIndicator'
 export type { HighlightIndicatorOptions } from './useHighlightIndicator'
 export { HIGHLIGHT_ATTR } from '@guillemservera/details-core/highlight-indicator'
+export type { HighlightStore } from '@guillemservera/details-core/highlight-indicator'
+export type { HighlightStoreState, Remeasure } from '../shared/highlight'